+++ internal/database/embeddings.sql.go
@@ -323,7 +323,7 @@ SELECT
     s.name,
     s.description,
     s.embedding_model,
-    (1 - (s.embedding <=> $1)) as similarity
+    (1 - (s.embedding <=> $1))::float8 as similarity
 FROM scry_quest.spells s
 WHERE s.embedding IS NOT NULL
 
@@ -335,7 +335,7 @@ SELECT
     b.name,
     CONCAT(b.size, ' ', b.type, ', ', b.alignment) as description,
     b.embedding_model,
-    (1 - (b.embedding <=> $1)) as similarity
+    (1 - (b.embedding <=> $1))::float8 as similarity
 FROM scry_quest.bestiary b
 WHERE b.embedding IS NOT NULL
 
@@ -347,7 +347,7 @@ SELECT
     c.name,
     c.description,
     c.embedding_model,
-    (1 - (c.embedding <=> $1)) as similarity
+    (1 - (c.embedding <=> $1))::float8 as similarity
 FROM scry_quest.classes c
 WHERE c.embedding IS NOT NULL
 
@@ -359,7 +359,7 @@ SELECT
     sp.name,
     sp.description,
     sp.embedding_model,
-    (1 - (sp.embedding <=> $1)) as similarity
+    (1 - (sp.embedding <=> $1))::float8 as similarity
 FROM scry_quest.species sp
 WHERE sp.embedding IS NOT NULL
 
@@ -378,7 +378,7 @@ type SearchAllContentByEmbeddingRow struct {
 	Name           string      `json:"name"`
 	Description    pgtype.Text `json:"description"`
 	EmbeddingModel pgtype.Text `json:"embedding_model"`
-	Similarity     int32       `json:"similarity"`
+	Similarity     float64     `json:"similarity"`
 }
 
 func (q *Queries) SearchAllContentByEmbedding(ctx context.Context, arg SearchAllContentByEmbeddingParams) ([]SearchAllContentByEmbeddingRow, error) {
