+++ internal/database/queries/species.sql
@@ -22,7 +22,7 @@ SELECT
     size,
     speed,
     embedding_model,
-    (1 - (embedding <=> $1)) as similarity
+    (1 - (embedding <=> $1))::float8 as similarity
 FROM scry_quest.species
 WHERE embedding IS NOT NULL
 ORDER BY embedding <=> $1
